+++ src/NuGetForUnity/Editor/PackageContentManager.cs
@@ -1,5 +1,5 @@
-﻿using System.IO;
-using System.Linq;
+﻿using System;
+using System.IO;
 using NugetForUnity.Configuration;
 using NugetForUnity.Helper;
 using NugetForUnity.Models;
@@ -96,6 +96,65 @@ namespace NugetForUnity
                 FileSystemHelper.DeleteDirectory(packageStreamingAssetsDirectory, true);
                 FileSystemHelper.DeleteFile($"{packageStreamingAssetsDirectory}.meta");
             }
+        }
+
+        internal static bool ShouldSkipUnpackingOnPath(string path, string packageId)
+        {
+            // skip a remnant .meta file that may exist from packages created by Unity
+            if (path.EndsWith($"{packageId}.nuspec.meta", StringComparison.Ordinal)) return true;
+
+            // skip directories & files that NuGet normally deletes
+            if (path.StartsWith("_rels", StringComparison.Ordinal) || path.Contains("/_rels/")) return true;
+            if (path.StartsWith("package", StringComparison.Ordinal) || path.Contains("/package/")) return true;
+            if (path.EndsWith($"{packageId}.nuspec", StringComparison.Ordinal)) return true;
+            if (path.EndsWith("[Content_Types].xml", StringComparison.Ordinal)) return true;
+
+            // Unity has no use for the build directory
+            if (path.StartsWith("build", StringComparison.Ordinal) || path.Contains("/build/")) return true;
+
+            // For now, skip src. We may use it later...
+            if (path.StartsWith("src", StringComparison.Ordinal) || path.Contains("/src/")) return true;
+
+            // Since we don't automatically fix up the runtime dll platforms, skip them until we improve support
+            // for this newer feature of nuget packages.
+            if (path.StartsWith("runtimes", StringComparison.Ordinal) || path.Contains("/runtimes/")) return true;
+
+            // Skip documentation folders since they sometimes have HTML docs with JavaScript, which Unity tried to parse as "UnityScript"
+            if (path.StartsWith("docs", StringComparison.Ordinal) || path.Contains("/docs/")) return true;
+
+            // Skip ref folder, as it is just used for compile-time reference and does not contain implementations.
+            // Leaving it results in "assembly loading" and "multiple pre-compiled assemblies with same name" errors
+            if (path.StartsWith("ref", StringComparison.Ordinal) || path.Contains("/ref/")) return true;
+
+            // Skip all PDB files since Unity uses Mono and requires MDB files, which causes it to output "missing MDB" errors
+            if (path.EndsWith(".pdb", StringComparison.Ordinal)) return true;
+
+            // Skip all folders that contain localization resource file
+            // Format of these entries is lib/<framework>/<language-code>/...
+            if (path.StartsWith("lib", StringComparison.Ordinal) || path.Contains("/lib/"))
+            {
+                var libSlashIndex = path.IndexOf("lib/", StringComparison.Ordinal) + 4;
+
+                var secondSlashIndex = path.IndexOf("/", libSlashIndex, StringComparison.Ordinal);
+                if (secondSlashIndex == -1)
+                {
+                    return false;
+                }
+
+                var thirdSlashIndex = path.IndexOf("/", secondSlashIndex + 1, StringComparison.Ordinal);
+                if (thirdSlashIndex == -1)
+                {
+                    return false;
+                }
+
+                var langLength = thirdSlashIndex - secondSlashIndex - 1;
+                if (langLength == 2 || langLength > 2 && path[secondSlashIndex + 3] == '-')
+                {
+                    return true;
+                }
+            }
+
+            return false;
         }
 
         private static string GetPackageInstallDirectory(INugetPackageIdentifier package)
